Remove underscore, append Protocol suffix, move to own files
