+++ Sources/SpelledPitch/PitchSpeller/Wetherfield/Data Structures/Graph/DirectedGraphProtocol.swift
@@ -1,0 +1,8 @@
+//
+//  DirectedGraphProtocol.swift
+//  SpelledPitch
+//
+//  Created by James Bean on 9/22/18.
+//
+
+import Foundation
